requirements.txt
pillow==8.0.1

# development
pytest==6.2.5
black==19.10b0
flake8==3.7.0
tqdm==4.62.3

